# docs/openapi.yaml
openapi: 3.0.3
info:
  title: GroceryGo / FreshMart API
  version: "1.0.0"
  description: >
    REST API for the FreshMart customer app and admin dashboard.
    Full endpoint reference: API_DOCUMENTATION.md. This spec covers the
    core resources structurally; extend it with the remaining routes
    listed there as needed (e.g. via `flask-smorest` or `apispec` if you
    want it auto-generated from the Flask blueprints going forward).
servers:
  - url: http://localhost:5000
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string, format: email }
        phone: { type: string, nullable: true }
        role: { type: string, enum: [admin, manager, staff, delivery, customer] }
        is_active: { type: boolean }
    Product:
      type: object
      properties:
        id: { type: integer }
        sku: { type: string }
        name: { type: string }
        price: { type: number, format: float }
        unit: { type: string }
        category: { type: string, nullable: true }
        stock_quantity: { type: integer }
        low_stock: { type: boolean }
    OrderItemInput:
      type: object
      required: [product_id, quantity]
      properties:
        product_id: { type: integer }
        quantity: { type: integer, minimum: 1 }
    Order:
      type: object
      properties:
        id: { type: integer }
        status: { type: string, enum: [pending, confirmed, preparing, out_for_delivery, delivered, cancelled] }
        total_amount: { type: number }
        delivery_address: { type: string }
    Error:
      type: object
      properties:
        error: { type: string }
paths:
  /api/auth/register:
    post:
      summary: Register a customer account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name: { type: string }
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                phone: { type: string }
                address: { type: string }
                city: { type: string }
                pincode: { type: string }
      responses:
        "201": { description: Created, content: { application/json: { schema: { type: object, properties: { data: { $ref: "#/components/schemas/User" } } } } } }
        "409": { description: Email already registered, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/auth/login:
    post:
      summary: Log in and receive JWT tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        "200": { description: OK }
        "401": { description: Invalid credentials, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/products:
    get:
      summary: List products
      security: []
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: category_id
          schema: { type: integer }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: per_page
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Product" }
    post:
      summary: Create a product (admin/manager)
      responses:
        "201": { description: Created }
        "401": { description: Unauthorized }
        "403": { description: Forbidden }
  /api/orders:
    get:
      summary: List orders visible to the current user
      responses:
        "200": { description: OK }
    post:
      summary: Place an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items, delivery_address]
              properties:
                items:
                  type: array
                  items: { $ref: "#/components/schemas/OrderItemInput" }
                delivery_address: { type: string }
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Order" }
        "409": { description: Insufficient stock, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/payments/orders/{orderId}/initiate:
    post:
      summary: Initiate payment for an order
      parameters:
        - in: path
          name: orderId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [method]
              properties:
                method: { type: string, enum: [upi, card, netbanking, cash] }
      responses:
        "201": { description: Payment initiated }
        "503": { description: Payment provider not configured, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }